Added some transformations for clips

+++ videotransforms/volume_transforms.py
@@ -2,7 +2,7 @@ import numpy as np
 from PIL import Image
 import torch
 
-from videotransforms.utils import images as imageutils
+from torch_videovision.videotransforms.utils import images as imageutils
 
 
 class ClipToTensor(object):
@@ -65,4 +65,4 @@ class ToTensor(object):
 
     def __call__(self, array):
         tensor = torch.from_numpy(array)
-        return tensor
+        return tensor.float()
